Fix bug in filter processing.
The FFT is twice as large as the signal we are processing.  This is partially to make the overlap and save work correctly.  We need to clear out the upper half of the signal buffer, otherwise when we use another FFT based filter, we'll use the entire thing and come up with corrupt data.

Signed-off-by: Jeremy McDermond <nh6z@nh6z.net>